+++ NoteMark/app/src/main/java/com/joshayoung/notemark/note/presentation/note_detail/NoteDetailAction.kt
@@ -5,5 +5,5 @@ import com.joshayoung.notemark.note.presentation.note_list.NoteListAction
 
 sealed interface NoteDetailAction {
     data object GoBack: NoteDetailAction
-    data class GoToEdit(val id: Int?) : NoteDetailAction
+    data class GoToEdit(val id: Long?) : NoteDetailAction
 }
