+++ NoteMark/app/src/main/java/com/joshayoung/notemark/note/presentation/note_list/NoteListAction.kt
@@ -6,6 +6,6 @@ import com.joshayoung.notemark.note.presentation.model.NoteUi
 sealed interface NoteListAction {
     data class OnDeleteClick(val noteUi: NoteUi): NoteListAction
     data object AddNoteClick: NoteListAction
-    data class GoToDetail(val id: Int) : NoteListAction
+    data class GoToDetail(val id: Long) : NoteListAction
     data object OnSettingsClick: NoteListAction
 }
